feat: comprehensive codebase review improvements
Refactor core module by extracting Group class from scene.py into
dedicated group.py, fix dimension ordering bugs, improve examples
compliance, enhance gsplats fitting/benchmark tooling, add Cholesky
dimension ops, and consolidate review reports into code_reviews/.

- Extract Group class from scene.py into core/group.py (SRP)
- Fix dimension ordering and add test_dim_order.py tests
- Delete in-tree REVIEW_REPORT.md files, move to code_reviews/
- Delete examples/COMPLIANCE_TODO.md (work completed)
- Update all examples with proper output dir and arbol usage
- Enhance CUDA benchmark.py with comparison and profiling modes
- Add Cholesky dimension operations and tests
- Add 4D multichannel gsplats demo
- Improve gsplat_data.py, trils.py, gpu_ops.py, decompose.py
- Update CLAUDE.md, .gitignore, core README/SPECIFICATIONS

+++ packages/luxar/examples/rainbow_sphere_spiral_example.py
@@ -78,7 +78,7 @@ def create_rainbow_colors(n_points: int) -> np.ndarray:
     g = np.sin(2 * np.pi * t + 2 * np.pi / 3) * 0.5 + 0.5
     b = np.sin(2 * np.pi * t + 4 * np.pi / 3) * 0.5 + 0.5
 
-    colors = np.column_stack([r, g, b])
+    colors = np.column_stack([r, g, b]).astype(np.float32)
 
     return colors
